requirements.txt
fastapi
uvicorn
pydantic
openai-agents
openai
qdrant-client
fastembed
arxiv
python-dotenv